Bug fix - Digest access auth variable nc is encoded in hexadecimal, not base ten.

+++ HTTPConnection.h
@@ -21,7 +21,7 @@
 	int numHeaderLines;
 	
 	NSString *nonce;
-	int lastNC;
+	long lastNC;
 	
 	NSObject<HTTPResponse> *httpResponse;
